keyboard revision
Added ' ̂ss' and took away from the 's' key ş as it was redundant.
Took away ç from the c key since it was redundant
Revised the readme.htm to include AZ and kjj.
Revised the name to be xinaliq everywhere.
Added the readme.htm to the kps file
Compiled the kpj file