Change the behaviour of ~ as explicit chandrakkala
Mozhi spec 2.0 was updated to make ~ as an explict virama key. So if a
user pressed ~, it shows his intent to preserve the chandrakkala
(virama). The keyboard now preserves this intent by adding a zwnj before
the joining consonant.

Adding this rule was tricky as we shouldn't spit out a zwnj when a user
type ~. This is because, adding so would cause the text to have hanging
(or widowed) zwnj. In order to prevent this from happening, I'm using a
deadkey to set a flag to remember that the ~ key was used for the
chandrakkala. When user type the next key, if it is a key that produces
another consonant, chillu or anusvara, a zwnj is placed between them.

*Potential Issue*
To my best knowledge a chillu can't appear after a chandrakkala. So, if
the user uses ~ and then a chillu producing keystroke, the current rule
will automatically output the base form of the chillu followed by a
virama. But if the user ever needed a chillu after a chandrakkala, the
explicit virama cancelling key, '_', have to be pressed.