fix(ci): dependencies for release packages
The dependencies for release packages were referring to other paths.
The pattern for the packages folder is to include the source kmn files
in the project so that they will be built along with the package.

However, the .kps files themselves still referred to the built kmx/js
in the original folders, rather than the expected ..\build\ path.

Given this change is a compile-phase-only change, there is no need to
update version numbers or make new releases.