Update welcome.htm
I am not allowed to create a folder for the help page, the welcome page was already merged into the master. I justed wanted the help page to be updated at 
https://help.keyman.com/keyboard/burushaski_girminas/1.0.2/burushaski_girminas.php. 
You can take the welcome page information to update the help page. Sorry I should have asked for it with he previous pull request.