[qom] Initial submission
This is the first version of the keyboard. We have done some testing with a limited number of users and it seems to be working as intended.  After it is publically available in Keyman's repository/website for download, we expect to continue developing it further by collecting much more feedback.