Requested Changes
Thank you very much, @LornaSIL, for all your great comments. And I feel a quick apology is in order, as some of the points you made were clearly stated in the documentation :)

I think I have now fixed all the issues you highlighted. Regarding the mobile layout, I made some changes but I'm purposefully leaving most of it as it is because we want this keyboard to be usable for both Spanish and Qom at the same time. In future releases, we will fine-tune it more and more.

Please, feel free to let me know if I missed anything!