chore: fixup special key styling
The Keyman Developer 14.0 compiler performs some additional checks on
the touch keyboard and identified a number of 'special keys' which do
not have the required `sp` styling. This PR fixes those, in order to
get the keyboards to compile with 14.0, but does not change the keyboard
versions, as this is probably not a significant enough change to warrant
updates.