fix: rewrite .kmp files that are incompatible with Keyman 8
Fixes #1527.

This simply rebuilds these kmp files to be compatible with Keyman
Desktop 8; there are no changes to the contents of the files. I will
manually upload these to the download server, because given the file
versions have not changed, they will not otherwise be updated.

I will attach a zip of the tool I used to rewrite these, for future
use if needed.