feat: xmllint .kps files
This adds xmllint checks for .kps file to the build. If xmllint is
not on the path, then the XMLLINT variable can be set to point to it;
if no xmllint can be found then .kps file structure validation will be
skipped (not a failure).

Found minor errors in one .kps file (apart from balochi_phonetic, which
is already resolved in a separate branch).

Relates to keymanapp/keyman#5464 and uses .xsd from
keymanapp/api.keyman.com#158. Do not merge until api.keyman.com PR is
approved (as we may need to propagate changes for the .xsd to this
branch).