feat(khmer_angkor): adds custom OSK font for better render
Relates to keymanapp/keyman#5986.

This swaps the on screen keyboard / touch layout keyboard font for a
custom font called Khmer Busra Kbd which resolves issues with shaping
across platforms by including custom PUA glyphs for any characters
which require shaping.

This is not yet complete:

1. Three characters are missing from the font: U+2248, U+17C7, U+17C8.
2. The font is not yet optimised for size.
3. I have only update the OSK and tablet, not yet the phone layout.

Still, this should be good for some testing, and will pick this up again
post-holidays.