chore: cleanup invalid touch layout files
These five files had some issues with invalid values:

1. finongan.keyman-touch-layout: invalid "pad" and "width" values
2. teggargrent_lat.keyman-touch-layout: invalid "pad" value
3. rac_yidgha.keyman-touch-layout: UTF-8 BOM mark
4. sil_cherokee_nation.keyman-touch-layout: UTF-8 BOM mark
5. sil_eastern_congo.keyman-touch-layout: invalid "width" value

Given these changes only impact source, I am not updating version
numbers and they do not need to be deployed.

I detected these changes using the draft schema for keyman-touch-layout
at keymanapp/keyman#6856. The schema is relatively loose, allowing
certain other values which should not really be permitted, so there may
be another round of tidy-up in the future.

Note: As the touch layout editor is currently reasonably lax on some of
this data, we probably need to tighten it up as well to avoid pain in
the future.