feat(sil_euro_latin): add Caps Lock and Start-of-Sentence detection
Caps Lock and Start-of-Sentence detection are new features for touch
devices in Keyman 15.0. The compiled keyboard remains backwardly
compatible with older versions of Keyman, as access to the new
functionality will just be ignored (this is known as progressive
enhancement). However, the keyboard must be built with the 15.0
compiler.

In order to test the new functionality, ensure you are running
Keyman 15.0.265 or later.