feat: regression builds
Adds scripts to do a full build of .kmx and .js files with a specific
version of the compiler. Puts results into output/<version>/ so that we
can easily run diff scripts against different versions.

Will first be used in the Keyman Developer CI build, which will download
latest version of compiler in the same tier, as well as using the build
within the Keyman Developer developer/release/<version> folder, which
should make for reasonably straightforward comparisons.

If the kmcomp.zip archive includes kmdecomp.exe, the script can also
decompile the .kmx keyboards back to .kmn, again in the
output/<version>/ folder, so that we can do textual comparisons of the
results.