chore: cleanup package files with missing metadata
The metadata in these packages was technically invalid, but does not
impact end users because they were desktop-only packages, and Keyman on
Windows at least did not make use of the metadata for legacy-style
packages. So we should not need a version bump for these. (Two of the
keyboards are also deprecated.)