chore: preprocess .keyboard_info, whitespace and encoding consistency
This commit makes no material changes to the content of the files. I
just loaded the .keyboard_info files into a node JSON parser and saved
them out with consistent formatting. This is groundwork for eventually
being able to remove the .keyboard_info files altogether, and is
intended to simplify the process of validating subsequent PRs which
*will* have changes to the content of the files.