chore: update help repo as a PR
This will open a pull request for keyboard layout documentation changes,
so then the changes will go through CI, which will avoid the current
issue where broken links in keyboards cause other PRs to fail CI.

It will also mean that we no longer push to master for keyboard help
publishing, which will be cleaner.

Initially, auto merge of these PRs will not happen. Once we are sure
that the process is working smoothly, we can setup a GH action on the
help repo to do this.