Taigi Telex keyboard second submission
I am sorry for that once I remove the number 2 inner test mark of my keyboard and do it in every files in the folder then compile it again, the new file will be totally can not be accepted although it still works on PC. I can not find the issue. Can you help me to find out the issue and remove the name? if you are not convenient, I accept that my keyboard name may contain number because it won't show for the users during usage. Because I really don't have time to redo my work for such a small issue that won't be shown during usage. 
Best regard,
cathaylab