The taigi telex version 2.0 "sdwxv" version
hello, this is my other variant of taigi telex. I adjusted a little bit about the keymap. I replaced the 'q' with 'd', and replaced 'd' with 's', It will be a totally different keyboard, and be another project. Is it the right way to submit may new keyboard? thank you.