Elymaic Keyboard
It's been a while since I did this so sorry if there's some mistakes. Please let me know if anything needs fixing. 
For the language, I couldn't find the code for Achaemenid Aramaic so I opted for the closest forms of Aramaic I could identify chronologically and geographically which people looking for an Elymaic script keyboard might expect to find it under.