fix(windows): registry settings defined in header shared
The imxconfig.cpp and KeymnIMX.cpp files had a different location
defined for accessing the configuration keys. This resulted in the
configuration being updated but not read correctly by the IMX dll.
Ideally the read and write functions should also be harmonised.
However, refactoring this code should soon become a yak shave.