Can you approve this keyboard? please
I understand the keyboard is super large. However, in every of my devices, this keyboard can work. If you think the keyboard is too large for a web keyboard. I can delete the web tick box.

I know this keyboard will use many of your resource, but today this is the only way to type old Hangul in a offline way. 

I know compiling this keyboard will spend you probably one hour. But finally it will be compiled in my computer.
Thank you so much.