feat: Added the letter v as an output
After descussion with other community members there is a strong desire
to use the letter v, to be added to their alphabet. This still needs
more conses from the comunity but we can give that has an option without
much trouble. I added v back in and moved ŋ to the /.