Update carambolfulbhe.kps
@sven-oly 
The path to the compiled files was wrong, so I fixed that. I added readme.htm, welcome.htm, and LICENSE.md into the list of files as well. We require LICENSE.md to be in the list of files and in the Details tab or our system will not build.